# PERF: improved performance of small multiindexes