--- a/pandas/core/util/hashing.py
+++ b/pandas/core/util/hashing.py
@@ -5,16 +5,13 @@ import itertools
 
 import numpy as np
 from pandas._libs import hashing
-from pandas._libs.lib import is_bool_array
 from pandas.core.dtypes.generic import (
     ABCMultiIndex,
     ABCIndexClass,
     ABCSeries,
     ABCDataFrame)
 from pandas.core.dtypes.common import (
-    is_categorical_dtype, is_numeric_dtype,
-    is_datetime64_dtype, is_timedelta64_dtype,
-    is_list_like)
+    is_categorical_dtype, is_list_like)
 
 # 16 byte long hashing key
 _default_hash_key = '0123456789123456'
@@ -136,7 +133,6 @@ def hash_tuples(vals, encoding='utf8', hash_key=None):
     -------
     ndarray of hashed values array
     """
-
     is_tuple = False
     if isinstance(vals, tuple):
         vals = [vals]
@@ -231,6 +227,7 @@ def hash_array(vals, encoding='utf8', hash_key=None, categorize=True):
 
     if not hasattr(vals, 'dtype'):
         raise TypeError("must pass a ndarray-like")
+    dtype = vals.dtype
 
     if hash_key is None:
         hash_key = _default_hash_key
@@ -238,22 +235,21 @@ def hash_array(vals, encoding='utf8', hash_key=None, categorize=True):
     # For categoricals, we hash the categories, then remap the codes to the
     # hash values. (This check is above the complex check so that we don't ask
     # numpy if categorical is a subdtype of complex, as it will choke.
-    if is_categorical_dtype(vals.dtype):
+    if is_categorical_dtype(dtype):
         return _hash_categorical(vals, encoding, hash_key)
 
     # we'll be working with everything as 64-bit values, so handle this
     # 128-bit value early
-    if np.issubdtype(vals.dtype, np.complex128):
+    elif np.issubdtype(dtype, np.complex128):
         return hash_array(vals.real) + 23 * hash_array(vals.imag)
 
     # First, turn whatever array this is into unsigned 64-bit ints, if we can
     # manage it.
-    if is_bool_array(vals):
+    elif isinstance(dtype, np.bool):
         vals = vals.astype('u8')
-    elif (is_datetime64_dtype(vals) or
-          is_timedelta64_dtype(vals)):
+    elif issubclass(dtype.type, (np.datetime64, np.timedelta64)):
         vals = vals.view('i8').astype('u8', copy=False)
-    elif (is_numeric_dtype(vals) and vals.dtype.itemsize <= 8):
+    elif issubclass(dtype.type, np.number) and dtype.itemsize <= 8:
         vals = vals.view('u{}'.format(vals.dtype.itemsize)).astype('u8')
     else:
         # With repeated values, its MUCH faster to categorize object dtypes,