@jorisvandenbossche 2017-05-11
@jreback Did you post the wrong timings? As you show it is actually slower ..