@jorisvandenbossche 2017-05-11
Or are the timings of the PR compared to 0.19.2 timings? (I assumed master or 0.20.1 timings) And in that case, with this PR we are still considerably slower compared to 0.19.2 (after first indexing operation and hashtable creation)

When I was testing it yesterday, I also assumed it was some trade-off on the size of the MI for the new hashtable based MultiIndexEngine, and the ObjectEngine. So I tried to find the good cut-off by varying the MI size, but I actually got systematically faster results with the Object engine, for MI's up to 1 million elements. 
So I now repeated that exercise but using this PR, fixing the engine to one of both implementations instead of letting it depend on the size:

```
In [16]: pd.MultiIndex.object_engine = False

In [17]: ts = []
    ...: 
    ...: for n in [3, 10, 30, 100, 300, 1000]:
    ...:     idx = pd.MultiIndex.from_product([np.arange(n), np.arange(n)])
    ...:     key = idx.values[0]
    ...:     t = %timeit -o idx.get_loc(key)
    ...:     ts.append(t)
    ...: 
    ...: 
100 loops, best of 3: 1.86 ms per loop
1000 loops, best of 3: 1.98 ms per loop
100 loops, best of 3: 2.28 ms per loop
100 loops, best of 3: 1.97 ms per loop
The slowest run took 5.58 times longer than the fastest. This could mean that an intermediate result is being cached.
100 loops, best of 3: 1.93 ms per loop
The slowest run took 89.39 times longer than the fastest. This could mean that an intermediate result is being cached.
1000 loops, best of 3: 2.23 ms per loop

In [18]: pd.MultiIndex.object_engine = True

In [19]: ts2 = []
    ...: 
    ...: for n in [3, 10, 30, 100, 300, 1000]:
    ...:     idx = pd.MultiIndex.from_product([np.arange(n), np.arange(n)])
    ...:     key = idx.values[0]
    ...:     t = %timeit -o idx.get_loc(key)
    ...:     ts2.append(t)
    ...: 
    ...: 
The slowest run took 52.39 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 5.07 µs per loop
The slowest run took 45.32 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 5.18 µs per loop
The slowest run took 65.73 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 5.21 µs per loop
The slowest run took 215.95 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 5.12 µs per loop
The slowest run took 2562.67 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 5.12 µs per loop
The slowest run took 21262.27 times longer than the fastest. This could mean that an intermediate result is being cached.
100000 loops, best of 3: 7.51 µs per loop
```

So the engine creation gets a lot slower (and that is important as well of course), but after that the actual indexing operation is still consistently faster when using the ObjectEngine, even for `len(self) > 1000`. 
The same for the example you showed above. And the strange thing is that in the original PR https://github.com/pandas-dev/pandas/pull/15245 you actually showed that the timing for that benchmark considerably improved.