@jorisvandenbossche 2017-05-11
Regarding that last remark about the asv benchmarks: the reason is that it does the indexing timings on a 'cold' index, so the result seems to be dominated by the engine creation rather than the actual `get_loc` operation. Maybe we should benchmark both.