@jreback 2017-05-11
@jorisvandenbossche pls read my impl notes. The timings are correct. Your measurement is incorrect.