@jorisvandenbossche 2017-05-11
I had read your comments in the classes, but I don't read an explanation for my comments above. So if you would like to clarify.

To be clear, I am not saying the ObjectEngine (the old implementation) is necessarily better (and thanks for this PR!). It is just that the trade-off between performance of index creation and of actual indexing operation, is a difficult one, because it depends on how many times you do such an indexing operation. So it will always be difficult to determine a good cut-off size to use one of both, because it depends on the usecase (apart from that, we should just choose a sensible one).

If you do many indexing operations (which is not most of the time not a good usage of course, but not always avoidable), the ObjectEngine is even for large arrays still a lot faster as the MultiIndexEngine (because in such case the faster `get_loc` outweighs the slower initial creation). 
If you do only a few indexing operation, the MultiIndexEngine, due to a much faster creation, can indeed be a big improvement.