@jreback 2017-05-11
@jorisvandenbossche your comments indicate you did not read my update.

> If you do many indexing operations (which is not most of the time not a good usage of course, but not always avoidable), the ObjectEngine is even for large arrays still a lot faster as the MultiIndexEngine (because in such case the faster get_loc outweighs the slower initial creation).

prove it